crypto: Move the verification cache into a separate module
